Update reproducer role to match new generator arch
Since ci_gen_kustomize_values is now handling "slightly" more
generations, we have to update the reproducer role to match the new
needs.

In order to avoid issues with `delegate_to`, facts and related
registered results, we now run most of the generation as a nested
playbook, launched directly via `ansible.builtin.command`. This ensures
we have the right context, and access to all of the needed parts.