Install telemetry_tempest_plugin from master branch
The version of the telemetry-tempest-plugin in the tempest container is
behind master.
The test in the tempest image require gnocchi, which is not included by
the telemetry operator. In order to test the telemetry-operator tempest
job, the newer version of the telemetry-tempest-plugin needs to be used.

In order to test this, the tempest role has been modified to update the
version of telemetery-tempest-plugin to be included.