Get zuul reproducer back on track
Since the introduction of networking_mapper in the reproducer role, the
zuul reproducer feature was broken.

This patch provides the needed update to re-enable the feature, by
remapping the networking_mapper environment output to the old
networking-info.yml format/file, allowing to re-run the job until we
switch the CI itself to the new networking_mapper usage.

This patch also provides the needed bits to properly support downstream
zuul and the RHEL ecosystem specificities. Most of the involved changes
are transparent to the user, since the tasks were already done on the
controller-0 - we had to extend them to the compute nodes.