Add cifmw_ssh_keytype and cifmw_ssh_keysize to create ssh keys
When ecdsa keys created with `community.crypto.openssh_keypair` are used
by paramiko, an error is raised (at least with paramiko versions 2.10
and 2.11).
They issue does not occur when they key is generated:
- with a non-empty comment
- with a size

This could affect tobiko and some tempest plugins (e.g.
whitebox-tempest-plugin, which connects to computes).

This patch modifies the ssh keys created by the tobiko role (to connect
to the VM instances created by tobiko) and the libvirt_manager role (to
connect to the VMs it creates) with size=521.

It uses global parameters cifmw_ssh_keytype and cifmw_ssh_keysize for
this.