Converge usage of the architecture repository
With #1045, we're seeing a new top-level parameter introduced in order
to point to the architecture repository.
We're also seeing some differences with the way scenarios are defined,
showing some more "defensive" approach toward the users, ensuring typos
should be avoided.

This patch should hopefully help limiting the differences to the strict
minimum, while correcting a couple of typo introduced in the first
iteration.

It also includes a couple of examples for the role, and some updates to
the reproducer role to match the new conventions.

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/1045