Clone all projects in the zuul reproducer
Clone all projects listed in the zuul inventory when reproducing a zuul
job, not just those coming from github. This is needed when reproducing
jobs defined in downstream gitlab that consume playbooks from internal
gerrit repos. If the internal gerrit repos are not in a depends-on, they
are not pulled and the playbooks are not found.