[test-operator] Introduce params for ssh-key and custom config files
In this patch [1] a new set of parameters were introduced to the
test-operatr. The 'SSHKeySecretName' enables the user to mount
custom private key via openshift secret and 'customOverwrite' gives
the user the ability to mount an arbitrary file with a content into
the tempest test pod.

This change exposes the newly introduced parameters via the
test-operator role.

[1] https://github.com/openstack-k8s-operators/test-operator/pull/39