Ensure support for newer CRC (>=2.32)
dnsmasq container was removed from newer CRC releases[1], leading to a
crash whenever we want to update its configuration.

This patch updates both reproducer and Zuul to ensure it's compatible
with older and newer releases.

Regarding Zuul, we dive slightly deeper in the way the resolvers where
configured in order to avoir the `crc-pre` service - this one is only
for CI.

[1] https://github.com/crc-org/snc/commit/35fce67a31e469f84e1047cc204771d12b5f6434