Make post-ceph ConfigMap values with cifmw_ceph_client
If the cifmw_ceph_client role is called with the variable
`cifmw_ceph_client_values_post_ceph_path_src` set to the
path of an HCI post Ceph values file [1], and with variable
`cifmw_ceph_client_service_values_post_ceph_path_src` set
to the path of a HCI post Ceph service values file [2],
then copy those files to produce a version of them with
the actual Ceph values substituted for the CHANGEME values.

The cifmw_ceph_client role already produces `k8s_ceph_secret.yaml`
with the same values. However, that file should not be "oc applied"
when the ci-framework deploys HCI VA, since those values should
instead be passed as input to a kustomization in the architecture
repo.

If the new variables are undefined, then the behavior of the
cifmw_ceph_client role is unchanged.

[1]
https://github.com/openstack-k8s-operators/architecture/blob/4af9a06c0a23f97bf73006e38931cf1e40597bc5/examples/va/hci/values.yaml
[2]
https://github.com/openstack-k8s-operators/architecture/blob/4af9a06c0a23f97bf73006e38931cf1e40597bc5/examples/va/hci/service-values.yaml

Signed-off-by: John Fulton <fulton@redhat.com>