OCP cluster as configured by dev-scripts uses volumes
We cannot just replace the "source" like it was done: libvirt is strict,
but also "dumb" in the way it handles a source with multiple conflicting
attributes.

This patch changes how OCP overlays are created - more precisely, the
location to be seen in the oooq_pool.

It also extracts the OCP XML manipulations in a way to make it easier to
loop.

We also ensure we clean the volumes on due time.