Return DLRN MD5 hash to the child job
Due to multiple promotions, the tcib job and child job gets
different current-podified hashes leading failure of pulling
containers in the EDPM deployment jobs.

Returning the dlrn md5 hash to the child job makes sure we
populate the same repo on the node.

Note: It also set the returned dlrn md5 hash to the repo-setup
to generate proper repos with passed dlrn hash.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>