Enable checkout from openstack ref flag in install_yamls
When enabling this flag in install_yamls, all repo cloning process
will consider the commit hash set in openstack-operator go.mod, which
will be be actually be installed when running the job.
To work across operator's repos, we need to also guarantee that
we set the ${OPERATOR_NAME}_REPO to the content of the PR, OR
provide an updated go.mod in openstack-operator that also has reference
to the PR content.
This patch modifies install_yamls role to include these additional vars
when running on zuul. It will search into zuul variables to identify
repositories paths from code changed in PRs and its depends-on.