Allow to properly scale nodes
Until now, the network-info.yml wasn't updated in case of re-deploy.
While it was OK with the first iteration of its generation, it doesn't
make sense anymore: we're gathering all the interfaces, always, for old
and new VM.

Removing the "force: false" will ensure we're properly populating the VM
interfaces, even in case of a scale-out, meaning a second run.