Remove cifmw_extras from the reproducer vars
The cifmw_extras is a variable we use to pass extra variables
to the nested ansible-playbooks we usually use in CI. That variable,
that has nothing to do with the framework itself, may contain jinja
templates, to shortcut the access to zuul project dirs for example.
That jinja, that the reproducer takes as a string, it's problematic when
playbooks inside the reproducer, like the autogenerated cifmw-bootstrap.yml,
tries to resolve it, cause variables, like zuul there, are not
available.
To avoid these issues, and based on the fact that cifmw_extras is not
a variable that the framework should care about out of our zuul CI,
we now filter it out.