Minor refactor in edpm_prepare and new hook for ceph scenarios
When deploying an HCI environment (together with edpm deploy), the
Control Plane resources are already ready and we don't have a chance of
kustomizing them with ceph backends. To enable some storage backend
with Ceph, we need to provide the Ceph FSID, which is available only
after edpm_deploy.
This change refactor edpm_prepare to split a set of tasks used to
kustomize and deploy the control plane again, if needed, after edpm
deploy. These tasks can be called in a hook (added in this PR as example)
to update the Control Plane resources.