Add tripleo controllers into adoption multinode source nodeset
As part of [1] this adds the tripleo_controller nodes into the
centos-9-multinode-rhel-9-2-crc-extracted-2.30-3xl nodeset used
by the multinode source adoption job. It sets the required
crc_ci_bootstrap_networking into the job vars as well as the
nodes into the rh-subscription group. We also move to the
cloud-rhel-9-tripleo nodeset for more memory.

https://issues.redhat.com/browse/OSPRH-5278

[1] https://issues.redhat.com/browse/OSPRH-3038