Change top-level parameters
The CI Framework shouldn't care about scenario type - VAs and DTs are
mostly the same. We want to care about defined automation only.

This patch removes most of the "va" from parameter names to be agnostic.

We also correct a bug introduced in the latest patches related to
automation: the missing shebang in the `deploy-architecture.sh` script
prevented using `ansible.builtin.command`.