[Tobiko] Introduce Tobiko Workflows
This patch enables to specify Tobiko workflow for the test-operator
role. This patch uses the newly introduced workflow section in the
Tobiko CR that can be used to specify several workflow steps. The
steps are executed sequentially.

This change ensures that we can set the workflow value and that the role
waits for the last job that executes the tobiko tests (by using the
workflowStep label)

Depends-On: https://github.com/openstack-k8s-operators/test-operator/pull/54