Adds a new single node HCI job
Adds a new single node HCI job and a job that uses the single
node HCI and configures control plane to use Ceph as backend.
Adds a pre_deploy kustomization for HCI environment, to set to
zero the number of GlanceAPIs replicas, since we can start it only
after deploying Ceph, in post_deploy phase.