[test-operator] Configure when test-operator asserts results
With this patch, when the test-operator runs tests from different test
frameworks, by default the results are asserted when all the execution
from all the tests frameworks is completed.
By configuring cifmw_test_operator_fail_fast to true, the role will
raise an error as soon as tests from a test framework fail.