Rework a bit conditions on playbook imports
With import_*, the catalog is built with all the content flattened, and
any conditions or vars/environment are applied on any of the flattened
content.

This means ansible will go through all of the tasks in the included
playbooks and "skip" them - instead of skipping the playbook import.

This patch should make things slightly more readable in the logs, with
an early end_play whenever we're able to apply it.

Note about the (pre|post)_deploy hooks:
we can't move them to the main `deploy-edpm.yml` playbook, because the
`06-deploy-edpm.yml` sub-play is directly imported in the adoption jobs.