Rework include/import usage in reproducer role
In order to make the whole reproducer slightly faster, a small rework in
the way we call roles has to be done.

`import_role` and `import_tasks` are embedding the content while building
the catalog, at the very beginning, and any `when` conditions are then
added to the subsequent tasks, leading to a lot of tasks being evaluated
and "skipped", clogging the logs and taking more time than necessary.

Switching "conditioned" imports to `include` are ensuring we don't face
that same behavior.

With this, we need to set the `cifmw_libvirt_manager_compute_amount`
parameter in the scenario files, since the libvirt_manager role isn't
*imported* anymore, but dynamically included.