add log collection for systemd services on edpm nodes
the edpm node contaienr and host service are now configured to log to the
journal instead of a file. while we do collect /var/log/messages which provides
a combined log that can be less easy to read without downloading it locally
when browsing ci logs.

This change add addtional log collection of each edpm* and *virt* systemd
service and outputs each to there own file for eaiser consumtion in the
the zuul web ui.