Allow users to inject their own custom hooks
This change empowers users with custom hooks injections during specific
stages in the Architecture deployment loop.

For instance, if a user wants to add their own hook as a "pre_stage",
for stage_id 2, they can create a custom parameter named:
`pre_stage_2_run_my_super_hook`. It will then be injected directly in
the existing pre_hook list of the stage.

If the user uses an index (0-9+ or a-z+) in the hook name, they can even ensure
the place their hook kicks in in the hook list:
`pre_stage_2_run_00_my_super_hook` has all the chances to be the very
first one, while `pre_stage_2_run_ZZ_my_super_hook` would be the last
one to run.