Correct generated scripts from reproducer role
`deploy-architecture.sh` was loading a generated parameter file, leading
to potential issues when a user is iterating deploys directly on
controller-0 and is passing parameters to the deployment: those
parameters would end in that generated file, and therefore would be
loaded on each run.
This patch corrects this, and ensure we're loading a "vanilla" generated
file, from the reproducer run itself.

`deploy-edpm.sh` script wasn't properly installed on the controller-0
due to a missing `delegate_to: controller-0` on the task. This wasn't
caught in CI, since we're using `zuul` user for both "hypervisor" and
"controller-0".
This patch adds the missing instruction, and the script lands in the
right place now.