Integrate virtualbmc in libvirt_manager
This patch allows to configure the virtualbmc service and inject all of
the created virtual machines in it.

Since we want to get only one instance of virtualbmc, even in a
multi-hypervisor layout, we have to be sure to properly target the one
hosting the OpenShift services, being OCP cluster or single CRC.

This patch also remove a duplicated call to the
libvirt_manager/clean_layout tasks from the devscripts role.

This patch also correct some indempotency issues in virtualbmc,
preventing a nice integration.

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/1284