Stop ceph playbooks as early as possible
With `import_playbook` ansible stills goes through all of the tasks
listed in the plays, but will `skip` them, leading to some time loss,
and potential issues with the expected inventory we have to consume.

We also ensure the HCI path of 06-deploy-edpm.yml isn't `imported` in
case we don't want to run it, switching to `include_*` instead. This
will ensure we don't have a long list of skipped tasks in the logs.

The default bahavior of the `ceph.yml` playbook isn't affected, meaning
we don't need to pass any parameter to NOT deploy it.