Add application_credential module
Create or delete a Keystone application credential.  When the secret
parameter is not set a secret will be generated and returned in the
response. Existing credentials cannot be modified so running this module
against an existing credential will result in it being deleted and
recreated. This needs to be taken into account when the secret is
generated, as the secret will change on each run of the module.

This will be removed when [1] lands. ci-bootstrap will use this module
to create application credentials for the duration of the job. This will
allow a sushy-tools container to have host cloud credentials which are
not the primary cloud credentials.

[1] https://review.opendev.org/c/openstack/ansible-collections-openstack/+/910463

Change-Id: I0ed86dc8785b0e9d10cc89cd9137a11d02d03945