Make cifmw_test_operator_*_config modifiable
When a consumer of the role runs the test_operator role with:

-e cifmw_test_operator_tempest_config=value

then the role can not update the value of the variable as extra vars
have higher precedence in ansible. This patch changes the structure
of the test_operator role so that we always work with a copy of this
variable.