Ensure we keep track of generated content
It may happen we want to generate twice the same values.yaml, but with
maybe different content - this is especially true in the VA automation,
for instance two stages share the same `src_file` and the same
`value.name`, but the output is different.

In such as case, we want to consume the exact same `values.yaml`
(untouched), the same CI generaged content, and maybe a different user
provided kustomization.

While it was 100% supported, we may lose track of the history of the
changes. This patch therefore ensures we:
- backup the files before override
- exclude those backup files from the final `values.yaml` generation