Donot generate nova migration key if exists
If we re-run the deploy-architecture.sh script. Then
Create nova migration keypair task fails with error:
```
Unable to read the key. The key is protected with a passphrase or broken.
```

In order to fix this issue, we donot need to regenerate the key
but to reuse it.

This pr checks for nova migration key if exists. Reuse the
same.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>