Establish framework for QE Validation jobs
We have differing requirements for QE validation jobs depending on the
team. This change aims to establish a framework by which each team will
be able to define their QE validation jobs. In doing so, the ci-framework
will be the extensible "tool to rule them all" from upstream CI to
downstream QE.

Accordingly, we establish a new role called `qe_validations`. Under
which we have tasks/<team> directories where validation jobs can be
defined. We can call these are part of a post-deployment execution in
the ci-framework. Essentially this would look like:

1. Normal deployment job is executed - as it would be today in ci
2. Post deployment will call a task from the qe_validations role to
   validate that certain conditions have been satisfied.
3. These qe_validation tasks should `assert` a state is satisfied.
   Assuming all assertions are satisfied, the job will be considered
   passed.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>