Extend and generate files based on IPMI data
In earlier patches, VirtualBMC was introduced to the CI Framework, with
an easy and small generated fact and subsequent file on the hypervisor.

This patch extends this fact in order to ensure we're exposing the
credentials, meaning we're now able to consume everything from some
other roles (in the same run).

In the reproducer role, we now take that virtualbmc fact, and extend it
a bit more, in order to add network interfaces information; we also
ensure key consistency by applying a "lower" mapping.