[Test-operator] Introduce Tempest workflow
Introduce the cifmw_test_operator_tempest_workflow variable which is
equivalent to the cifmw_test_operator_tobiko_workflow variable.

This option enables the specification of a value for the workflow field
in the Tempest CR. Using the workflow, the user can specify multiple
executions of Tempest in a specific order.

Depends-On: https://github.com/openstack-k8s-operators/test-operator/pull/67