Add UEFI support for the virtual machines
Until now, only BIOS mode was supported for the VM managed by the
libvirt_manager role.

From now on, we add support for UEFI, provided the base image supports
it of course (note: this is NOT the case for the official
CentOS-stream-9 images!).

A quick check/automation allows to detect if the image "is supposed to
support" UEFI based on cifmw_repo_setup_os_release parameter value. RHEL
images do provide UEFI, so let's just use it.