Allow to set a specific port for VBMC daemon
This port wasn't documented, and of course may conflict if you have a
pre-existing VBMC service running on the host (especially in devscripts
scenario).

This patch adds a minimalistic VirtualBMC configuration file to the
container, allowing to set a specific port for the main daemon.

The patch also updates the va-hci.yml scenario file to allow a smoother
deployment.