Integrate approve_csr module as part of the workflow
In this change set, a check for calling approve_csr is introduced
so that we can avoid when doing a fresh deployment. The checks
identifies if the cert expiration file is older than a day. If
yes, then approve_csr module is executed.

In this change, the number of attempts to be made for authenticating
against the cluster via the login subcommand is reduced to 10. This
is done as the login check happens after stablizing the cluster.

Signed-off-by: Pragadeeswaran Sathyanarayanan <psathyan@redhat.com>