Ensure script properly reports failure in reproducer
Until now, if the deploy-architecture.sh was failing, it was silent:
ansible didn't know about that failure, and just reported green.

Adding the `set -e` seems to ensure we're reporting the state correctly.

This flag is therefore added to the deploy-architecture.sh and
deploy-edpm.sh scripts (even though the second isn't called by the
reproducer - let's just be consistent).