Try to get a faster re-run for libvirt_manager
By registering `overlay` state, we should be able to avoid a long list
of tasks that are useless against an already created VM.

We also slightly amend the way we generated the SSH configuration for
the ProxyJump in order to allow running multiple "labs" on multiple
hypervisor, while avoiding overriding the whole ssh configuration and
losing access to the VMs.

With this patch, in case you're deploying two labs on two distinct
machines, you'll be able to reach either of them by using, for instance,
`ssh controller-0.hypervisor-1` or `ssh compute-0.hypervisor-2`.