Ensure we know all the hosts beforehand
Since #1343, ci_local_storage is using plain ansible to create the
persistent storage on the OpenShift node (in this case, CRC).

This means it's using SSH - and since the job never connects to the
nested CRC instance, it's failing due to the unknown remote host key.

Having ssh-keyscan populating the known_hosts should prevent this issue.