Ensure we're using the right VM layout
In the `libvirt_manager/deploy_layout.yml` we select the "layout" we
want to deploy. It generates a cached fact named `_layout`, and should
be using this one exclusively.

In parallel, the `reproducer` role should also use the correct layout,
as soon as the fact is available.

For the records, in the "zuul reproducer" path, the layout may be
updated by ansible in order to reflect the correct amount of computes,
using an internal parameter named
`cifmw_libvirt_manager_configuration_gen` - `libvirt_manager` then
selects either that parameter if exists, or defaults to the standard
`cifmw_libvirt_manager_configuration`, creating that `_layout` internal
fact.

This patch reconcile the parameter usage.