Support replacing of links with a copy of the source file.
kustomize does not support resources that are symbolic links. Also,
ansible kustomize filter does not support relaxing of loading
restrictions. Due to this limitations, it is necessary to make a
copy of the requested resource files.

This feature is required in cases the kustomization file in DT
or libraries need to be subset.

Signed-off-by: Pragadeeswaran Sathyanarayanan <psathyan@redhat.com>