Support executing of commands as part of hooks.
There are cases wherein we would need to execute a command apart
from the validation that is being done. This could be a post or
pre step that is required for the kustomize to progress.

Example: A node needs to be applied a specific label so that
lvm-iscsi can make progress.

Signed-off-by: Pragadeeswaran Sathyanarayanan <psathyan@redhat.com>