[Test-operator] Fix mounting of ssh key
Currently, the private key specified using the default value for
the SSHKeySecretName is the content of ~/.ssh/id_cifw. The file is mounted
to the test pod as a secret that is generated using the test-operator
role. The content of the file is loaded using lookup function that
strips the newline at the end of the file by default. This fix ensures
that the newline is kept omitting the newline breaks the usage of the
key with ssh.