Allow longer names for hooks related content
The ci_script module is truncating at 30 chars, and keep full words. The
`truncate` filter in ansible allows to do the same, so let's converge.

This patch also removes a now useless `ansible.builtin.debug` tasks:
ci_script module outputs the log location directly, so we don't need to
do it in a dedicated task anymore.