Correct issues due to cached facts and k8s core module
First issue was related to cached facts. It ended up with a weird
situation where the controller-0 was using facts that weren't
up-to-date, especially visible in a scale-out scenario (adding some
computes).

The second issue is related to `kubernetes.core.k8s` module not updating
the resources:
we would need to first check for the resource existence before deciding
if we would use `present` or `patched` for the state parameter.

Using plain `oc apply` with a valid `change_when` allows to get rid of
that issue.