Allow to inject custom wait_conditions during stages
This feature is needed, especially in HCI scenario, when we want to
scale-out the deployment:
in order to do so, we have to create a new OpenStackDataPlaneDeployment
resource, with a new name to trigger the deployment of the new
compute(s).

In such a use-case, using [1] and the following parameters, we're able
to wait for the right resource to be ready:

```YAML
cifmw_architecture_user_kustomize:
  stage_2:
    edpm-values:
      data:
        deployment:
          name: edpm-scale-out
cifmw_architecture_wait_condition:
  stage_2:
    - >-
      oc -n openstack wait
      osdpd edpm-scale-out --for condition=Ready
      --timeout=1500s
```

In this example, the first snippet will inject a user kustomization in
the ci_gen_kustomize_values role in order to override the resource name.
The second snippet will then inject a new wait_condition to the existing
one.
When the deployment hits the wait_conditions loop, it will pass over the
first one (since it's already done), and run the second condition until
it ends.

[1] https://github.com/openstack-k8s-operators/architecture/pull/158