Ensure we have an undying virtualbmc
It seems vbmcd process dies on its own without much log after a period
of time.
Overriding the CMD of the container with an infinite loop corrects the
issue for the time being, and would allow to access the container for
debugging in case of a crash.

The container logs as shown by `podman logs <container>` don't help,
since we just learn that vbmcd "stopped", without error. We may be able
to find something in the container itself, if vbmcd is writing logs on
its own.