Move again NFS server after EDPM deploy
Otherwise the target EDPM node has no full networking setup,
which means the choosen VLAN where to expose

This means that any usage of NFS should be moved to the post_deploy
phase, possibly reconfiguring an existing control plane.

This mirror the similar mechanism used to configure and use
Ceph on EDPM nodes.