Add new feature in the reproducer: molecule jobs
This patch enables molecule job reproducer, at least for the ones
configured in the ci-framework (not sure if molecule is running in any
other places).

This patch allows the reproducer role to detect if we're reproducing a
molecule job or a "standard" deployment; it will update the layout
accordingly (it needs #1439 for the `amount: 0` support), and create the
needed playbook/script to run the whole stack accordingly.

Several changes were needed in the reproducer role in order to ensure
we're consistent with the data we fetch from Zuul, and some tasks had to
move in dedicated files to make things easier to read, as well as
leveraging the `include_tasks` (to avoid endless list of `skipping` log
lines).

This change also needs #1440 to ensure we're able to pass a few
parameters down in the job playbook.

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/1440
Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/1439