Wait for secret creation in install_openstack_ca
Wait for a few seconds to check if the combined-ca-bundle secret is
created, using retries. In molecule tests, it seems that the secret takes a bit to be
created, so the tests were failing because we were getting the secret
too quickly. The task has ignore_errors: true to match the previous
behaviour of ignore-errors=true.