Update CRC configuration in reproducer
Since newer CRC introduce an OVS connection we can't really manage (the
configuration is reset upon reboot), we can't override the dns
configuration.

Since it uses the hypervisor as a DNS server, the best thing is to
update the /etc/hosts there to ensure CRC will know about itself...

The main issue here is, the hypervisor is first in the list of resolver,
leading to potential issues if we don't update the CRC record in
/etc/hosts.

The `ansible.builtin.blockinfile` is crafted in a way to ensure it
overrides the block created by CRC itself during its deployment.