Allow to push secrets on controller-0
By leveraging `cifmw_manage_secrets_reproducer_secrets` parameter, we
are now able to push secrets from the ansible controller (your laptop,
for instance) onto controller-0.

If the destination is a relative path (just a file name for instance) it
will be pushed into cifmw_basedir/secrets - else you can pass an
absolute path. Keep in mind it's on the controller-0, meaning the user
is `zuul`.