[openshift_login] Only slurp install-yamls-params when update is needed
The install-yamls-params.yml param file is always slurped.
The task that uses this slurped file is only run conditionally.
Add the slurp to a conditional block with the file-update task.

The bootstrap playbook, which runs install_yamls and creates the file
can skip running install_yamls when the bootstrap tag is skipped.
This causes errors when the infra playbook is run and expects the params
file to exist so it can be read.