Add dataplane hotfix QE job
This PR adds validation tests to ensure functionality of the
dataplane-operator to rollout hotfixes to the EDPM nodes.

It does this by:
1. Updating the image used by `edpm_nova_compute_image` in the OpenStackDataPlaneNodeSet
2. Rolling out those changes with a new OpenStackDataPlaneDeployment
3. Waiting for the new OpenStackDataPlaneDeployment to be complete
4. Gathering the currently running image from the node
5. Asserting the provided hotfix image matches the running image on the
   node for the nova_compute container.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>

Use ci_script instead of command/shell

This change switches use of command and shell modules to instead use
the ci-framework specific ci_script module. This allows for a consistent gathering of artifacts
between the ci execution, and the QE validation being executed.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>