Correct ssh configuration cleaning
While the hypervisor name was added to the "mark" on the laptop, we
don't have it when we update the config on the hypervisor itself.

This patch corrects the wrong cleaning.

In order to ensure environment are in a right state, it also ensures
older configuration "marker" are removed from the laptop. This block may
be removed in a few weeks.