Fix push_code and zuul_vars vs zuul
With the addition of the molecule job reproducer, we had to take
parameters one level up in the environment. Those parameters were at the
same level as `zuul` - so we created a new `zuul_vars`, containing `zuul` itself.

This patch now checks for both `zuul` and `zuul_vars`, so that we can
match in CI runtime and in a job reproducer.