Switch from "import" to "include"
The tasks were using `import_role` with a `when` condition block.

In such a case, the `when` conditions are applied to all of the imported
content, and ansible will have to consider each task and apply the `when`
block, meaning:
- loss of time (it takes a few seconds for each task to consider)
- lot of useless "skipped" log entries

With `include_role`, the `when` block is considered only for this task,
and if the condition is `false`, then nothing more happens, speeding up
and cleaning the logs.