Move generated YAML files to more appropriate location
the reproducer-variables.yml and openshift-environment.yml were in
zuul's home directory on controller-0.

This location is outside of any "ci-framework known location". It means
it won't get caught by current tools, while they provide really useful
information for debugging, or understanding of the environment.

For backward compatibility, symlinks to the old location exist