Delete all the resources of openshift-marketplace
https://github.com/ansible-collections/kubernetes.core/pull/517 added
delete_all: true to delete all the resources.

It needs to be added in fix openshift-marketplace task files
to make sure all pods are deleted otherwise
fix_openshift_marketplace.yml taskfile does not work as expected.

Note: It moves to oc commands to be make sure all the pods are in
Running state.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>