Group common parameters for Architecture deployment
A series of parameters are mandatory in all cases. Let's avoid
duplication and group them in a single file.

That single file is then loaded if and only if we set the
`cifmw_architecture_scenario` parameter to a non-empty string.

As noted in the va-common.yml file, ANY parameter there can be
overridden by a parameter file, or even using `-e foo=bar` directly.

This should make "creating new VA or DT" process slightly easier for
everyone, while providing a unified way to fix all the scenarios in case
of new parameter/feature/others.