add missing ssh-known-hosts service to hci_prepare
nova requires that a ssh known hosts file is created on
all hosts however this service was missing from the
nodeset created in the hci_prepare role.

The hci_prepare role replaces the default set of services
used by the nodeset with a custom set, this change adds
ssh-known-hosts to that list in the same location as the
default nodeset sample.