Extract podman related bits in its own role
Since we'll see more podman containers in the near future, let's get a
deidcated role to group the needed actions.

This new `podman` role installs the package, and enable session
lingering if requestes (defaults `true`).

We can also override the package list to install in case we want more of
podman-related content, such as buildah, skopeo and so on.